Renamed glide module to imageloader
