+++ collect_app/build.gradle
@@ -249,7 +249,7 @@ dependencies {
     implementation project(':projects')
     implementation project(':location')
     implementation project(':geo')
-    implementation project(':glide')
+    implementation project(':imageloader')
     implementation project(':errors')
     implementation project(':upgrade')
     implementation project(':permissions')
