+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -130,8 +130,8 @@ import org.odk.collect.async.Scheduler;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.audiorecorder.recording.AudioRecorderFactory;
 import org.odk.collect.forms.FormsRepository;
-import org.odk.collect.glide.ImageLoader;
-import org.odk.collect.glide.GlideImageLoader;
+import org.odk.collect.imageloader.ImageLoader;
+import org.odk.collect.imageloader.GlideImageLoader;
 import org.odk.collect.location.LocationClient;
 import org.odk.collect.location.LocationClientProvider;
 import org.odk.collect.permissions.PermissionsChecker;
