+++ settings.gradle
@@ -19,4 +19,4 @@ include ':errors'
 include ':externalapp'
 include ':upgrade'
 include ':permissions'
-include ':glide'
+include ':imageloader'
